# [SPARK-31398][SQL] Fix perf regression of loading dates before 1582 year by non-vectorized ORC reader