@MaxGekk 2020-04-10
> Can you check the benchmark numbers with Spark 2.4? Just want to see how much perf regression we have in 3.0 after this patch.

@cloud-fan To have comparable results, need to port:
1. NoOp datasource 
2. Changes in Benchmark framework to save results to files