@cloud-fan 2020-04-10
1. we can use `df.queryExecution.toRdd.foreach`
2. we don't need to commit the benchmark result in 2.4, just need to take a look at the numbers and put it in PR description.